requirements.txt
streamlit
torch
diffusers
transformers
Pillow
